feat[#183]: ShortsCommentsByTimeResponseDto의 time의 타입을 long으로 변환

+++ src/main/java/org/highfive/backend/shorts/dto/response/ShortsCommentsByTimeResponseDto.java
@@ -1,7 +1,7 @@
 package org.highfive.backend.shorts.dto.response;
 
 public record ShortsCommentsByTimeResponseDto(
-       int time,
+       Long time,
        String username,
        String profileUrl,
        String comment,
